make compatible with customizepkg
The customizepkg tool searches for lines similar to the regular
expression "pkg.*=" to evaluate.

A pkgname is needed to determine which customization should be applied.
It is first assigned to be the package group name. The assignment is
captured by the regular expression. Then, it is unset again, which is
not captured, and rebuilt to be the actual extension names, which is
again captured, but without any effect because the variables composed
are without any value.

Lastly, introducing a line break when defining the pkgdescs keeps
customizepkg from failing while trying to execute the grep subshell
expression.